refactor: constrain sse ping configuration to integer seconds #114

+++ src/codex_a2a_server/config.py
@@ -96,8 +96,8 @@ class Settings(BaseSettings):
         default=1.0,
         alias="A2A_CANCEL_ABORT_TIMEOUT_SECONDS",
     )
-    a2a_stream_sse_ping_seconds: float = Field(
-        default=10.0,
+    a2a_stream_sse_ping_seconds: int = Field(
+        default=10,
         alias="A2A_STREAM_SSE_PING_SECONDS",
     )
     a2a_stream_idle_diagnostic_seconds: float = Field(
@@ -118,7 +118,7 @@ class Settings(BaseSettings):
 
     @field_validator("a2a_stream_sse_ping_seconds")
     @classmethod
-    def validate_stream_sse_ping_seconds(cls, value: float) -> float:
+    def validate_stream_sse_ping_seconds(cls, value: int) -> int:
         if value <= 0:
             raise ValueError("A2A_STREAM_SSE_PING_SECONDS must be > 0")
         return value
